Move counter example to its own directory to make space for more examples
