+++ example/src/CountExample/Contexts/EvenOrOddContext.ts
@@ -1,8 +1,10 @@
 import { createSuperContext } from "react-super-context";
 import { useCounter } from "./CounterContext";
 
-export const [evenOrOdd, useEvenOrOdd] = createSuperContext(() => {
+const [evenOrOddContext, useEvenOrOdd] = createSuperContext(() => {
   const { count } = useCounter();
 
   return count % 2 === 0 ? "even" : "odd";
 });
+
+export { evenOrOddContext, useEvenOrOdd };
