+++ example/src/CountExample/Contexts/LoggingContext.ts
@@ -3,7 +3,7 @@ import { useCounter } from "./CounterContext";
 import { useEvenOrOdd } from "./EvenOrOddContext";
 import { createSuperContext } from "react-super-context";
 
-export const [logging] = createSuperContext(() => {
+const [loggingContext] = createSuperContext(() => {
   const { count } = useCounter();
   const evenOrOdd = useEvenOrOdd();
 
@@ -11,3 +11,5 @@ export const [logging] = createSuperContext(() => {
     console.log(`The current count is ${count} which is ${evenOrOdd}`);
   }, [count, evenOrOdd]);
 });
+
+export { loggingContext };
